+++ www/docs/typedoc/add-frontmatter.mjs
@@ -1,7 +1,7 @@
 // @ts-check
 
 import { MarkdownPageEvent } from 'typedoc-plugin-markdown'
-import { entryFileName, modulesFileName } from '../constants.mjs'
+import { entryFileName, modulesFileName } from './constants.mjs'
 
 /** @param {import('typedoc-plugin-markdown').MarkdownApplication} app */
 export function load(app) {
